# backend/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: BASELINE Field Operations Intelligence API
  version: 1.0.0
  description: |-
    BASELINE captures operational truth through field visits and derives intelligence from that truth.
    Every endpoint documents its required role, request/response shapes, and applicable audit log action.

    **Role hierarchy**: Admin > Lead > UL (Unit Leader / field worker)
    **Auth**: Bearer token (Azure Entra JWT) on all non-public endpoints.
    **Labor safety**: No endpoint exposes per-worker performance metrics or identifiers
    to Lead/UL roles. `captured_by_oid` is encrypted and Admin-access-only.
  contact:
    name: BASELINE Engineering
servers:
  - url: http://localhost:3001/api
    description: Local development
  - url: /api
    description: Production (relative)
...
components:
  securitySchemes:
    AzureAD:
      type: oauth2
      description: 'Azure Active Directory (Entra) OAuth2. Pass the resulting access token as `Authorization: Bearer <token>`.'
      flows:
        authorizationCode:
          authorizationUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/authorize
          tokenUrl: https://login.microsoftonline.com/{tenantId}/oauth2/v2.0/token
          scopes:
            openid: OpenID Connect
            profile: User profile
            email: Email address
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message
          example: Resource not found
  responses:
    BadRequest:
      description: Request validation failed — missing or invalid field
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: pool_id is required
    Unauthorized:
      description: Authentication required or Bearer token is invalid/expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Unauthorized
    Forbidden:
      description: Caller lacks the required role for this endpoint
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Forbidden
...
    InternalError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: Internal server error
paths:
...
  /admin/tenant/asset-types:
    get:
      summary: List asset types for a tenant
      description: >-
        Returns all asset types configured for the specified org. Org is supplied via X-Org-Id header or ?org_id= query
        param.
      tags:
        - Tenant
      security:
        - AzureAD: []
      x-required-roles:
        - Admin
      parameters:
        - in: header
          name: X-Org-Id
          schema:
            type: integer
          description: Org ID (alternative to ?org_id=)
          example: 1
        - in: query
          name: org_id
          schema:
            type: integer
          description: Org ID (alternative to X-Org-Id header)
          example: 1
      responses:
        '200':
          description: List of asset types
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_types:
                    type: array
                    items:
                      type: object
              example:
                asset_types:
                  - id: 1
                    type_key: transit_stop
                    display_name: Transit Stop
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Create an asset type for a tenant
      description: Registers a new asset type for the org. Used during tenant onboarding.
      tags:
        - Tenant
      security:
        - AzureAD: []
      x-required-roles:
        - Admin
      parameters:
        - in: header
          name: X-Org-Id
          schema:
            type: integer
          description: Org ID
          example: 1
        - in: query
          name: org_id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type_key
                - display_name
              properties:
                type_key:
                  type: string
                  description: Unique slug for this asset type
                  example: transit_stop
                display_name:
                  type: string
                  example: Transit Stop
                description:
                  type: string
            example:
              type_key: transit_stop
              display_name: Transit Stop
              description: Bus shelter or transit stop
      responses:
        '201':
          description: Asset type created
          content:
            application/json:
              schema:
                type: object
                properties:
                  asset_type:
                    type: object
              example:
                asset_type:
                  id: 1
                  type_key: transit_stop
                  display_name: Transit Stop
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenant/observation-types:
    get:
      summary: List observation types for an asset type
      description: Returns all observation types configured for a given asset type within the org.
      tags:
        - Tenant
      security:
        - AzureAD: []
      x-required-roles:
        - Admin
      parameters:
        - in: header
          name: X-Org-Id
          schema:
            type: integer
          example: 1
        - in: query
          name: org_id
          schema:
            type: integer
        - in: query
          name: asset_type_id
          required: true
          schema:
            type: integer
          description: Asset type ID to list observation types for
          example: 1
      responses:
        '200':
          description: List of observation types
          content:
            application/json:
              schema:
                type: object
                properties:
                  observation_types:
                    type: array
                    items:
                      type: object
              example:
                observation_types:
                  - id: 1
                    observation_key: cleanliness
                    display_name: Cleanliness
                    value_type: state
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      summary: Upsert observation types for an asset type
      description: Creates or updates observation type definitions for a given asset type. Used during tenant onboarding.
      tags:
        - Tenant
      security:
        - AzureAD: []
      x-required-roles:
        - Admin
      parameters:
        - in: header
          name: X-Org-Id
          schema:
            type: integer
          example: 1
        - in: query
          name: org_id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - asset_type_id
                - types
              properties:
                asset_type_id:
                  type: integer
                  example: 1
                types:
                  type: array
                  items:
                    type: object
                    required:
                      - observation_key
                      - display_name
                      - value_type
                    properties:
                      observation_key:
                        type: string
                      display_name:
                        type: string
                      value_type:
                        type: string
                        enum:
                          - state
                          - numeric
                          - boolean
                      valid_values:
                        type: array
                        items:
                          type: string
                      is_required:
                        type: boolean
                      sort_order:
                        type: integer
            example:
              asset_type_id: 1
              types:
                - observation_key: cleanliness
                  display_name: Cleanliness
                  value_type: state
                  valid_values:
                    - clean
                    - dirty
                    - hazardous
                  is_required: true
                  sort_order: 1
      responses:
        '200':
          description: Observation types upserted
          content:
            application/json:
              schema:
                type: object
                properties:
                  upserted:
                    type: array
                    items:
                      type: object
              example:
                upserted:
                  - id: 1
                    observation_key: cleanliness
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /admin/tenant/seed-assets:
    post:
      summary: Seed asset inventory from a CSV upload
      description: >
        Accepts a multipart CSV file with required columns: external_id, display_name, lat, lon. Additional columns
        become entries in the asset attributes JSON. Used for bulk onboarding of asset inventories (e.g., importing stop
        data from a GIS export).
      tags:
        - Tenant
      security:
        - AzureAD: []
      x-required-roles:
        - Admin
      parameters:
        - in: header
          name: X-Org-Id
          schema:
            type: integer
          example: 1
        - in: query
          name: org_id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
                - asset_type_id
              properties:
                file:
                  type: string
                  format: binary
                  description: CSV file with columns external_id, display_name, lat, lon, and optional extras
                asset_type_id:
                  type: integer
                  description: Asset type to assign to all seeded assets
                  example: 1
      responses:
        '200':
          description: Assets seeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  rows_received:
                    type: integer
                  inserted:
                    type: integer
                  updated:
                    type: integer
              example:
                rows_received: 450
                inserted: 420
                updated: 30
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'